refactor: split one common media vm into subclasses
